Fix withLoggedInUser
Well, assuming I did. Will actually have to get around to using
it further down the line and seeing whether it does what I want.

There were some problems with userId returning Maybe a userid, which
in turn changed go's entire signature. In the end I added a redirect
if userid were to be a Nothing. Note however that this should never
ever be possible.